+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 6.9.3
 - Fix JS Markdown to HTML blockquote in warehouse.js
 - Fix deprecated create_function() in Security.php
 - Fix Do not allow N/A for Period select: allow N/A when not new in Courses.php
+- Fix Take Attendance link in Portal.php
 
 Changes in 6.9.2
 ----------------